Add region option to configure AWS region of container
This region option will fall back to the specified or default profile's region. This required bringing in profiles into the runner, which actually made things cleaner.
However, this removed the ability to specify an sts_client in the runner.rb constructor.

I also changed the CLI option from --no_creds to --no_credentials.